Transferred to new branch. Added double barrel shotgun weapon.
Double-barrel shotgun.
Very effective at close range sending anyone right next to you into crit in one shot, and has the possiblity of removing an arm or leg, or even both if you're lucky.
It has a special caliber requirement. It cannot use buckshot or similar ammo. It has its own ammotype.

It has a unique method of reloading. Requires you to break the shotgun open before you can reload it. This requried the implementation of a new child obejct of gun_parent, known /obj/item/gun/hardreload. At the moment this object is made specifically with the double-barrel shotgun weapon in mind. This can be extended and modified to create a more general specification that would allow for weapons to have more complex reloading systems.

The overall niche this gun fits is as a weapon that's primarily useful in close range. It can be pocketed for discreteness, and is very powerful when used in its element. However, it's balanced by being complex to reload, having a small ammo capacity (one shot), and not being very effective beyond short range. It's a weapon you would want to pair with a reliable stun weapon.